+++ lib/omq/routing/dealer.rb
@@ -27,7 +27,6 @@ module OMQ
       # @param connection [Connection]
       #
       def connection_added(connection)
-        @connections << connection
         add_fair_recv_connection(connection)
         add_round_robin_send_connection(connection)
       end